refactor(domain): 로또 일치 판단 로직, 대상 도메인 내부로 메소드 이동
당첨 번호와 로또 번호 일치 개수 판단 메소드, WinningNumber 내부로 이동
보너스 번호와 로또 번호 일치 여부 판단 메소드, BonusNumber 내부로 이동
이에 따라, WinningChecker 클래스 제거
이는 당첨 숫자와 보너스 번호 값 자체를 외부 클래스로 제공하지 않기 위함이다.
이에 따라 테스트도 각 도메인테스트 내부로 단위테스트로 수정하여 이동한다.
