+++ src/test/java/lotto/domain/WinningNumberTest.java
@@ -1,11 +1,13 @@
 package lotto.domain;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.Arrays;
 import java.util.List;
 import lotto.exception.InvalidWinningNumberException;
 import lotto.exception.LottoError;
+import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
@@ -69,5 +71,15 @@ class WinningNumberTest {
         assertThatThrownBy(() -> new WinningNumber(numbers))
                 .isInstanceOf(InvalidWinningNumberException.class)
                 .hasMessage(LottoError.WINNING_NUMBER_DUPLICATED.getMessage());
+    }
+
+    @Test
+    void 로또_번호가_당첨_번호와_3개_일치하는_경우_일치_개수는_3이_반환된다() {
+        WinningNumber winningNumber = new WinningNumber(List.of(1, 2, 3, 4, 5, 6));
+        Lotto lotto = Lotto.from(List.of(1, 2, 3, 10, 11, 12));
+
+        int result = winningNumber.countMatchesWith(lotto);
+
+        assertThat(result).isEqualTo(3);
     }
 }
